requirements.txt
Flask-Marshmallow
marshmallow
marshmallow-sqlalchemy
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-JWT-Extended
Flask-Marshmallow
marshmallow-sqlalchemy
Flask-CORS
python-dotenv
authlib requests
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary